CLN: Move rolling helper functions to where they are used
